+++ ukf/unscented_kalman_filter.cc
@@ -96,7 +96,7 @@ void UnscentedKalmanFilter::Constrain(ukfVectorType& x, const ukfMatrixType& W)
 // matrix version
 void UnscentedKalmanFilter::Constrain(ukfMatrixType& localX, const ukfMatrixType& localW)
 {
-  const unsigned int numCols = localX.cols();
+  const unsigned int numCols = static_cast<unsigned int>(localX.cols());
 
   for( unsigned int i = 0; i < numCols; ++i )
     {
