+++ ukf/utilities.cc
@@ -28,7 +28,7 @@ ukfPrecisionType l2fa(ukfPrecisionType l1, ukfPrecisionType l2, ukfPrecisionType
 ukfPrecisionType s2ga(const ukfMatrixType& signal)
 {
 
-  int n = signal.rows();
+  int n = static_cast<int>(signal.rows());
 
   assert(signal.cols() == 1);
 
@@ -55,7 +55,7 @@ ukfPrecisionType s2ga(const ukfMatrixType& signal)
 ukfPrecisionType s2adc(const ukfMatrixType& signal)
 {
 
-  int n = signal.rows();
+  int n = static_cast<int>(signal.rows());
 
   assert(signal.cols() == 1);
 
@@ -72,7 +72,7 @@ ukfPrecisionType s2adc(const ukfMatrixType& signal)
 
 ukfPrecisionType curve_radius(const stdVec_t& fiber)
 {
-  int length = fiber.size();
+  size_t length = fiber.size();
 
   if( length < 3 )
     {
